Address Layer 2a code review (helper relocation, O(1) fallback, semantic alignment)
Three review findings consolidated:

* Move build_globalid_xml_lookup and pair_peep_with_xml from
  extract_temporal_data to estleg_common (canonical home of
  iter_peep_files). Layer 2b's preamble scanner is the third consumer
  and would have made this churn worse later. The helpers are
  re-exported from extract_temporal_data for one cycle so existing
  test imports keep working.

* Replace O(N) 'in xml_lookup.values()' scan with O(1) set lookup
  in extract_temporal_data. Mirrors the pattern already used in
  generate_amendment_history. ~100x speedup on this comparison.

* Align _fallback_hits semantics across the 3 pipelines that share
  these helpers. Legal-concepts was tracking 'peep lacks globalId'
  instead of 'XML resolved via slug fallback'. Now consistent.
